chore: improve tsconfig for incremental builds
