+++ packages/backend/src/routers/inviteLinksRouter.ts
@@ -1,12 +1,12 @@
 import { CreateInviteLink, getObjectValue } from '@lightdash/common';
-import express from 'express';
+import express, { type Router } from 'express';
 import {
     allowApiKeyAuthentication,
     isAuthenticated,
     unauthorisedInDemo,
 } from '../controllers/authentication';
 
-export const inviteLinksRouter = express.Router();
+export const inviteLinksRouter: Router = express.Router();
 
 inviteLinksRouter.get(
     '/:inviteLinkCode',
