+++ packages/backend/src/routers/jobsRouter.ts
@@ -1,10 +1,10 @@
-import express from 'express';
+import express, { type Router } from 'express';
 import {
     allowApiKeyAuthentication,
     isAuthenticated,
 } from '../controllers/authentication';
 
-export const jobsRouter = express.Router({ mergeParams: true });
+export const jobsRouter: Router = express.Router({ mergeParams: true });
 
 jobsRouter.get('/', isAuthenticated, async (req, res, next) => {
     next('Not implemented');
