+++ packages/backend/src/routers/passwordResetLinksRouter.ts
@@ -1,7 +1,7 @@
-import express from 'express';
+import express, { type Router } from 'express';
 import { unauthorisedInDemo } from '../controllers/authentication';
 
-export const passwordResetLinksRouter = express.Router();
+export const passwordResetLinksRouter: Router = express.Router();
 
 passwordResetLinksRouter.get(
     '/:code',
