+++ packages/backend/src/routers/userRouter.ts
@@ -1,10 +1,10 @@
-import express from 'express';
+import express, { type Router } from 'express';
 import {
     isAuthenticated,
     unauthorisedInDemo,
 } from '../controllers/authentication';
 
-export const userRouter = express.Router();
+export const userRouter: Router = express.Router();
 
 userRouter.patch(
     '/me',
