+++ packages/common/src/index.ts
@@ -1,21 +1,9 @@
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
 import { z } from 'zod';
-import {
-    type ApiUnusedContent,
-    type ApiUserActivityDownloadCsv,
-    type UnusedContent,
-    type UnusedContentItem,
-    type UserActivity,
-    type ViewStatistics,
-} from './types/analytics';
-import {
-    type Dashboard,
-    type DashboardAvailableFilters,
-    type DashboardBasicDetails,
-    type DashboardSummary,
-} from './types/dashboard';
-import { type Explore, type SummaryExplore } from './types/explore';
+import { type AnyType } from './types/any';
+import { LightdashMode } from './types/api';
+import { type Explore } from './types/explore';
 import {
     DimensionType,
     friendlyName,
@@ -31,169 +19,21 @@ import {
     type Field,
     type FieldId,
     type FilterableDimension,
-    type FilterableField,
     type ItemsMap,
     type Metric,
     type TableCalculation,
 } from './types/field';
-import {
-    type AdditionalMetric,
-    type MetricQuery,
-    type QueryWarning,
-} from './types/metricQuery';
-import {
-    OrganizationMemberRole,
-    type ApiOrganizationMemberProfiles,
-    type OrganizationMemberProfile,
-} from './types/organizationMemberProfile';
-import {
-    type CreatePersonalAccessToken,
-    type PersonalAccessToken,
-} from './types/personalAccessToken';
-import { type ProjectMemberProfile } from './types/projectMemberProfile';
-import {
-    type ApiCalculateSubtotalsResponse,
-    type ApiCalculateTotalResponse,
-    type ChartHistory,
-    type ChartVersion,
-    type SavedChart,
-    type Series,
-} from './types/savedCharts';
-import { type SearchResults } from './types/search';
-import { type ShareUrl } from './types/share';
-import { type SlackSettings } from './types/slackSettings';
-import { type ApiCreateTagResponse } from './types/tags';
-
-import {
-    type ApiCreateComment,
-    type ApiDeleteComment,
-    type ApiGetComments,
-} from './types/api/comments';
-import { type Email } from './types/api/email';
-import { type ApiSuccessEmpty } from './types/api/success';
-import { type ApiGetChangeResponse } from './types/changeset';
-import { type DbtExposure } from './types/dbt';
-import { type EmailStatusExpiring } from './types/email';
-import { type FieldValueSearchResult } from './types/fieldMatch';
-import { type DashboardFilters } from './types/filter';
-import {
-    type GitIntegrationConfiguration,
-    type GitRepo,
-    type PullRequestCreated,
-} from './types/gitIntegration';
-import {
-    type DeleteOpenIdentity,
-    type OpenIdIdentitySummary,
-} from './types/openIdIdentity';
-import {
-    type AllowedEmailDomains,
-    type OnboardingStatus,
-    type Organization,
-    type OrganizationProject,
-    type UpdateAllowedEmailDomains,
-} from './types/organization';
-import { type ApiTogglePinnedItem, type PinnedItems } from './types/pinning';
-import { type ProjectGroupAccess } from './types/projectGroupAccess';
-import { type ProjectMemberRole } from './types/projectMemberRole';
+import { type AdditionalMetric, type MetricQuery } from './types/metricQuery';
+import { OrganizationMemberRole } from './types/organizationMemberProfile';
 import {
     DbtProjectType,
     ProjectType,
-    type CreateWarehouseCredentials,
     type DbtProjectConfig,
-    type Project,
-    type WarehouseCredentials,
 } from './types/projects';
-import { type MostPopularAndRecentlyUpdated } from './types/resourceViewItem';
-import { type ResultColumns, type ResultRow } from './types/results';
-import {
-    type ApiJobScheduledResponse,
-    type ApiJobStatusResponse,
-    type ApiSchedulersResponse,
-    type SchedulerAndTargets,
-    type SchedulerJobStatus,
-} from './types/scheduler';
-import { type ApiSlackChannelsResponse } from './types/slack';
-import { SpaceMemberRole, type CreateSpace, type Space } from './types/space';
-import { type ApiSshKeyPairResponse } from './types/SshKeyPair';
-import { type TableBase } from './types/table';
-import {
-    type LightdashUser,
-    type LoginOptions,
-    type UserAllowedOrganization,
-} from './types/user';
-import { type UserWarehouseCredentials } from './types/userWarehouseCredentials';
-import { type ValidationResponse } from './types/validation';
-
-import type {
-    ApiAiAgentAdminConversationsResponse,
-    ApiAiAgentArtifactResponse,
-    ApiAiAgentEvaluationResponse,
-    ApiAiAgentEvaluationRunResponse,
-    ApiAiAgentEvaluationRunResultsResponse,
-    ApiAiAgentEvaluationRunSummaryListResponse,
-    ApiAiAgentEvaluationSummaryListResponse,
-    ApiAiAgentThreadCreateResponse,
-    ApiAiAgentThreadGenerateTitleResponse,
-    ApiAiAgentThreadMessageCreateResponse,
-    ApiAiAgentThreadMessageVizQueryResponse,
-    ApiAiAgentThreadMessageVizResponse,
-    ApiAiAgentThreadResponse,
-    ApiAiAgentThreadSummaryListResponse,
-    ApiAiOrganizationSettingsResponse,
-    ApiAppendInstructionResponse,
-    ApiCreateEvaluationResponse,
-    ApiGetUserAgentPreferencesResponse,
-    ApiUpdateAiOrganizationSettingsResponse,
-    ApiUpdateUserAgentPreferencesResponse,
-    DecodedEmbed,
-    EmbedUrl,
-} from './ee';
-import { type AnyType } from './types/any';
-import {
-    type ApiGetProjectParametersListResults,
-    type ApiGetProjectParametersResults,
-} from './types/api/parameters';
-import { type ApiGetSpotlightTableConfig } from './types/api/spotlight';
-import { type Account } from './types/auth';
-import {
-    type ApiCatalogAnalyticsResults,
-    type ApiCatalogMetadataResults,
-    type ApiGetMetricsTree,
-    type ApiMetricsCatalog,
-} from './types/catalog';
-import {
-    type ApiChartAsCodeListResponse,
-    type ApiChartAsCodeUpsertResponse,
-    type ApiDashboardAsCodeListResponse,
-} from './types/coder';
-import {
-    type ApiChartContentResponse,
-    type ApiContentResponse,
-} from './types/content';
-import type { ApiGroupListResponse } from './types/groups';
-import type {
-    ApiMetricsExplorerQueryResults,
-    ApiMetricsExplorerTotalResults,
-} from './types/metricsExplorer';
-import type { ResultsPaginationMetadata } from './types/paginateResults';
-import { type ParametersValuesMap } from './types/parameters';
-import { type PivotConfiguration } from './types/pivot';
-import { type ApiPromotionChangesResponse } from './types/promotion';
-import { type QueryHistoryStatus } from './types/queryHistory';
-import { type ApiRenameFieldsResponse } from './types/rename';
-import { type SchedulerWithLogs } from './types/schedulerLog';
-import {
-    type ApiCreateSqlChart,
-    type ApiCreateVirtualView,
-    type ApiGithubDbtWritePreview,
-    type ApiSqlChart,
-    type ApiSqlRunnerJobStatusResponse,
-    type ApiUpdateSqlChart,
-    type GroupByColumn,
-    type SortBy,
-} from './types/sqlRunner';
+import { type ResultRow } from './types/results';
+import { type Series } from './types/savedCharts';
+import { SpaceMemberRole, type CreateSpace } from './types/space';
 import { TimeFrames } from './types/timeFrames';
-import { type ApiWarehouseTableFields } from './types/warehouse';
 import { convertAdditionalMetric } from './utils/additionalMetrics';
 import { getFields } from './utils/fields';
 import { formatItemValue } from './utils/formatting';
@@ -202,9 +42,17 @@ import { getOrganizationNameSchema } from './utils/organization';
 import type { PivotValuesColumn } from './visualizations/types';
 
 dayjs.extend(utc);
-export * from './authorization/index';
+export * from './authorization/buildAccountHelpers';
+export {
+    defineUserAbility,
+    getUserAbilityBuilder,
+    JWT_HEADER_NAME,
+} from './authorization/index';
+export * from './authorization/jwtAbility';
+export * from './authorization/parseAccount';
 export * from './authorization/roleToScopeMapping';
 export * from './authorization/scopes';
+export * from './authorization/serviceAccountAbility';
 export * from './authorization/types';
 export * from './compiler/exploreCompiler';
 export * from './compiler/filtersCompiler';
@@ -215,8 +63,11 @@ export * from './constants/sessionStorageKeys';
 export * from './constants/sqlRunner';
 export { default as DbtSchemaEditor } from './dbt/DbtSchemaEditor/DbtSchemaEditor';
 export * from './dbt/validation';
-export * from './ee/index';
-export * from './pivot';
+export * from './ee';
+export * from './pivot/derivePivotConfigFromChart';
+export * from './pivot/pivotConfig';
+export * from './pivot/pivotQueryResults';
+export * from './pivot/utils';
 export { default as lightdashDbtYamlSchema } from './schemas/json/lightdash-dbt-2.0.json';
 export { default as lightdashProjectConfigSchema } from './schemas/json/lightdash-project-config-1.0.json';
 export * from './templating/template';
@@ -325,7 +176,10 @@ export * from './utils/fields';
 export * from './utils/filters';
 export * from './utils/formatting';
 export * from './utils/github';
-export * from './utils/i18n';
+export * from './utils/i18n/chartAsCode';
+export * from './utils/i18n/dashboardAsCode';
+export * from './utils/i18n/merge';
+export * from './utils/i18n/types';
 export * from './utils/item';
 export * from './utils/loadLightdashProjectConfig';
 export * from './utils/metricsExplorer';
@@ -557,215 +411,6 @@ export const snakeCaseName = (text: string): string =>
 
 export const hasSpecialCharacters = (text: string) => /[^a-zA-Z ]/g.test(text);
 
-export type CacheMetadata = {
-    cacheUpdatedTime?: Date;
-    cacheExpiresAt?: Date;
-    cacheKey?: string;
-    cacheHit: boolean;
-};
-
-export type ApiQueryResults = {
-    metricQuery: MetricQuery;
-    cacheMetadata: CacheMetadata;
-    rows: ResultRow[];
-    fields: ItemsMap;
-};
-
-type ApiExecuteAsyncQueryResultsCommon = {
-    queryUuid: string;
-    cacheMetadata: CacheMetadata;
-    parameterReferences: string[]; // params needed for query to run
-    usedParametersValues: ParametersValuesMap; // params values used
-};
-
-export type ApiExecuteAsyncMetricQueryResults =
-    ApiExecuteAsyncQueryResultsCommon & {
-        metricQuery: MetricQuery;
-        fields: ItemsMap;
-        warnings: QueryWarning[];
-    };
-
-export type ApiExecuteAsyncDashboardChartQueryResults =
-    ApiExecuteAsyncQueryResultsCommon & {
-        metricQuery: MetricQuery;
-        fields: ItemsMap;
-        appliedDashboardFilters: DashboardFilters;
-    };
-
-export type ApiExecuteAsyncSqlQueryResults =
-    ApiExecuteAsyncQueryResultsCommon & {
-        // leaving empty for now
-    };
-
-export type ApiExecuteAsyncDashboardSqlChartQueryResults =
-    ApiExecuteAsyncQueryResultsCommon & {
-        appliedDashboardFilters: DashboardFilters;
-    };
-
-export type ReadyQueryResultsPage = ResultsPaginationMetadata<ResultRow> & {
-    queryUuid: string;
-    columns: ResultColumns;
-    rows: ResultRow[];
-    initialQueryExecutionMs: number;
-    resultsPageExecutionMs: number;
-    status: QueryHistoryStatus.READY;
-    pivotDetails: {
-        // Unlimited total column count, this is used to display a warning to the user in the frontend when the number of columns is over MAX_PIVOT_COLUMN_LIMIT
-        totalColumnCount: number | null;
-        indexColumn: PivotConfiguration['indexColumn'] | undefined;
-        valuesColumns: PivotValuesColumn[];
-        groupByColumns: GroupByColumn[] | undefined;
-        sortBy: SortBy | undefined;
-        originalColumns: ResultColumns;
-    } | null;
-};
-
-export type ApiGetAsyncQueryResults =
-    | ReadyQueryResultsPage
-    | {
-          status: QueryHistoryStatus.PENDING | QueryHistoryStatus.CANCELLED;
-          queryUuid: string;
-      }
-    | {
-          status: QueryHistoryStatus.ERROR;
-          queryUuid: string;
-          error: string | null;
-      };
-
-export type ApiDownloadAsyncQueryResults = {
-    fileUrl: string;
-};
-
-export type ApiDownloadAsyncQueryResultsAsCsv = {
-    fileUrl: string;
-    truncated: boolean;
-};
-
-export type ApiDownloadAsyncQueryResultsAsXlsx = {
-    fileUrl: string;
-    truncated: boolean;
-};
-
-export type ApiChartAndResults = {
-    chart: SavedChart;
-    explore: Explore;
-    appliedDashboardFilters: DashboardFilters | undefined;
-    metricQuery: MetricQuery;
-    cacheMetadata: CacheMetadata;
-    rows: ResultRow[];
-    fields: ItemsMap;
-};
-
-export type ApiSqlQueryResults = {
-    fields: Record<string, { type: DimensionType }>;
-    rows: Record<string, unknown>[];
-};
-
-export type ApiScheduledDownloadCsv = {
-    jobId: string;
-};
-export type ApiDownloadCsv = {
-    url: string;
-    status: SchedulerJobStatus;
-    truncated: boolean;
-};
-
-export type ProjectCatalog = {
-    [database: string]: {
-        [schema: string]: {
-            [table: string]: Pick<TableBase, 'description' | 'sqlTable'>;
-        };
-    };
-};
-
-export enum TableSelectionType {
-    ALL = 'ALL',
-    WITH_TAGS = 'WITH_TAGS',
-    WITH_NAMES = 'WITH_NAMES',
-}
-
-export type TablesConfiguration = {
-    tableSelection: {
-        type: TableSelectionType;
-        value: string[] | null;
-    };
-};
-
-export type CreateProjectMember = {
-    email: string;
-    role: ProjectMemberRole;
-    sendEmail: boolean;
-};
-
-export type UpdateProjectMember = {
-    role: ProjectMemberRole;
-};
-
-export type UpdateMetadata = {
-    upstreamProjectUuid?: string | null; // null means we unset this value
-};
-export type ApiCompiledQueryResults = {
-    query: string;
-    parameterReferences: string[];
-};
-
-export type ApiExploresResults = SummaryExplore[];
-
-export type ApiExploreResults = Omit<Explore, 'unfilteredTables'>;
-
-export type ApiStatusResults = 'loading' | 'ready' | 'error';
-
-export type ApiRefreshResults = {
-    jobUuid: string;
-};
-
-export type ApiCreatePreviewResults = {
-    projectUuid: string;
-    compileJobUuid: string;
-};
-
-export type ApiJobStartedResults = {
-    jobUuid: string;
-};
-
-export type ApiCreateUserTokenResults = {
-    token: string;
-    expiresAt: Date;
-};
-
-export type ActivateUser = {
-    firstName: string;
-    lastName: string;
-    password: string;
-};
-
-export type CreateUserArgs = {
-    firstName: string;
-    lastName: string;
-    email: Email;
-    password: string;
-};
-
-export type CreateUserWithRole = {
-    firstName: string;
-    lastName: string;
-    email: Email;
-    password?: string;
-    role: OrganizationMemberRole;
-};
-
-export type ActivateUserWithInviteCode = ActivateUser & {
-    inviteCode: string;
-};
-
-export type RegisterOrActivateUser =
-    | ActivateUserWithInviteCode
-    | CreateUserArgs;
-
-export const hasInviteCode = (
-    data: RegisterOrActivateUser,
-): data is ActivateUserWithInviteCode => 'inviteCode' in data;
-
 export const CompleteUserSchema = z.object({
     organizationName: getOrganizationNameSchema().optional(),
     jobTitle: z.string().min(0),
@@ -776,245 +421,6 @@ export const CompleteUserSchema = z.object({
 
 export type CompleteUserArgs = z.infer<typeof CompleteUserSchema>;
 
-export type UpdateUserArgs = {
-    firstName: string;
-    lastName: string;
-    email: string;
-    isMarketingOptedIn: boolean;
-    isTrackingAnonymized: boolean;
-    isSetupComplete: boolean;
-    isActive: boolean;
-};
-
-export type PasswordResetLink = {
-    expiresAt: Date;
-    code: string;
-    email: string;
-    url: string;
-    isExpired: boolean;
-};
-
-export type CreatePasswordResetLink = Pick<PasswordResetLink, 'email'>;
-
-export type PasswordReset = {
-    code: string;
-    newPassword: string;
-};
-
-export type ApiHealthResults = HealthState;
-export type InviteLink = {
-    expiresAt: Date;
-    inviteCode: string;
-    inviteUrl: string;
-    organizationUuid: string;
-    userUuid: string;
-    email: string;
-};
-export type CreateInviteLink = Pick<InviteLink, 'expiresAt' | 'email'> & {
-    email: string;
-    role?: OrganizationMemberRole;
-};
-
-export type ApiCreateProjectResults = {
-    project: Project;
-    hasContentCopy: boolean;
-};
-
-export type ProjectSavedChartStatus = boolean;
-
-export type ApiFlashResults = Record<string, string[]>;
-
-export type ApiAiDashboardSummaryResponse = {
-    status: 'ok';
-    results: DashboardSummary;
-};
-
-export type ApiAiGetDashboardSummaryResponse = {
-    status: 'ok';
-    results: DashboardSummary;
-};
-
-export type ApiAiGenerateCustomVizResponse = {
-    status: 'ok';
-    results: string;
-};
-
-type ApiResults =
-    | ApiQueryResults
-    | ApiSqlQueryResults
-    | ApiCompiledQueryResults
-    | ApiExploresResults
-    | ApiExploreResults
-    | ApiStatusResults
-    | ApiRefreshResults
-    | ApiCreatePreviewResults
-    | ApiHealthResults
-    | Organization
-    | LightdashUser
-    | LoginOptions
-    | SavedChart
-    | SavedChart[]
-    | Space[]
-    | InviteLink
-    | OrganizationProject[]
-    | Project
-    | WarehouseCredentials
-    | OrganizationMemberProfile[]
-    | ProjectCatalog
-    | TablesConfiguration
-    | Dashboard
-    | DashboardBasicDetails[]
-    | OnboardingStatus
-    | Dashboard[]
-    | DeleteOpenIdentity
-    | ApiFlashResults
-    | Record<OpenIdIdentitySummary['issuerType'], OpenIdIdentitySummary[]>
-    | FilterableField[]
-    | DashboardAvailableFilters
-    | ProjectSavedChartStatus
-    | null
-    | Array<unknown>
-    | ApiJobStartedResults
-    | ApiCreateUserTokenResults
-    | CreatePersonalAccessToken
-    | PersonalAccessToken
-    | ProjectMemberProfile[]
-    | ProjectGroupAccess
-    | SearchResults
-    | Space
-    | ShareUrl
-    | SlackSettings
-    | ApiSlackChannelsResponse['results']
-    | UserActivity
-    | UnusedContent
-    | UnusedContentItem
-    | ApiUnusedContent
-    | SchedulerAndTargets
-    | SchedulerAndTargets[]
-    | FieldValueSearchResult
-    | ApiDownloadCsv
-    | AllowedEmailDomains
-    | UpdateAllowedEmailDomains
-    | UserAllowedOrganization[]
-    | EmailStatusExpiring
-    | ApiScheduledDownloadCsv
-    | PinnedItems
-    | ViewStatistics
-    | SchedulerWithLogs
-    | ValidationResponse[]
-    | ChartHistory
-    | ChartVersion
-    | EmbedUrl
-    | DecodedEmbed
-    | Array<GitRepo>
-    | PullRequestCreated
-    | GitIntegrationConfiguration
-    | UserWarehouseCredentials
-    | ApiJobStatusResponse['results']
-    | ApiJobScheduledResponse['results']
-    | ApiSshKeyPairResponse['results']
-    | MostPopularAndRecentlyUpdated
-    | ApiCalculateTotalResponse['results']
-    | Record<string, DbtExposure>
-    | ApiCreateComment['results']
-    | ApiGetComments['results']
-    | ApiDeleteComment
-    | ApiSuccessEmpty
-    | ApiCreateProjectResults
-    | ApiAiDashboardSummaryResponse['results']
-    | ApiAiGetDashboardSummaryResponse['results']
-    | ApiCatalogMetadataResults
-    | ApiCatalogAnalyticsResults
-    | ApiPromotionChangesResponse['results']
-    | ApiWarehouseTableFields['results']
-    | ApiTogglePinnedItem['results']
-    | ApiOrganizationMemberProfiles['results']
-    | ApiSqlChart['results']
-    | ApiCreateSqlChart['results']
-    | ApiUpdateSqlChart['results']
-    | ApiContentResponse['results']
-    | ApiChartContentResponse['results']
-    | ApiSqlRunnerJobStatusResponse['results']
-    | ApiCreateVirtualView['results']
-    | ApiGithubDbtWritePreview['results']
-    | ApiMetricsCatalog['results']
-    | ApiMetricsExplorerQueryResults['results']
-    | ApiGroupListResponse['results']
-    | ApiCreateTagResponse['results']
-    | ApiChartAsCodeListResponse['results']
-    | ApiDashboardAsCodeListResponse['results']
-    | ApiChartAsCodeUpsertResponse['results']
-    | ApiGetMetricsTree['results']
-    | ApiMetricsExplorerTotalResults['results']
-    | ApiGetSpotlightTableConfig['results']
-    | ApiCalculateSubtotalsResponse['results']
-    | ApiExecuteAsyncSqlQueryResults
-    | ApiExecuteAsyncDashboardSqlChartQueryResults
-    | ApiExecuteAsyncMetricQueryResults
-    | ApiExecuteAsyncDashboardChartQueryResults
-    | ApiGetAsyncQueryResults
-    | ApiSchedulersResponse['results']
-    | ApiUserActivityDownloadCsv['results']
-    | ApiRenameFieldsResponse['results']
-    | ApiDownloadAsyncQueryResults
-    | ApiDownloadAsyncQueryResultsAsXlsx
-    | ApiAiAgentThreadResponse['results']
-    | ApiAiAgentThreadMessageVizResponse['results']
-    | ApiAiAgentThreadMessageVizQueryResponse['results']
-    | ApiUpdateUserAgentPreferencesResponse['results']
-    | ApiGetUserAgentPreferencesResponse[`results`]
-    | ApiGetProjectParametersResults
-    | ApiGetProjectParametersListResults
-    | ApiAiAgentThreadCreateResponse['results']
-    | ApiAiAgentThreadMessageCreateResponse['results']
-    | ApiAiAgentArtifactResponse['results']
-    | ApiAiAgentThreadGenerateTitleResponse['results']
-    | ApiAiAgentThreadSummaryListResponse['results']
-    | Account
-    | ApiAiAgentAdminConversationsResponse['results']
-    | ApiAiAgentEvaluationSummaryListResponse['results']
-    | ApiAiAgentEvaluationResponse['results']
-    | ApiAiAgentEvaluationRunResponse['results']
-    | ApiAiAgentEvaluationRunSummaryListResponse['results']
-    | ApiAiAgentEvaluationRunResultsResponse['results']
-    | ApiCreateEvaluationResponse['results']
-    | ApiAppendInstructionResponse['results']
-    | ApiGetChangeResponse['results']
-    | ApiAiOrganizationSettingsResponse['results']
-    | ApiUpdateAiOrganizationSettingsResponse['results'];
-
-export type ApiResponse<T extends ApiResults = ApiResults> = {
-    status: 'ok';
-    results: T;
-};
-
-export type ApiErrorDetail = {
-    name: string;
-    statusCode: number;
-    message: string;
-    data: { [key: string]: string };
-    sentryTraceId?: string;
-    sentryEventId?: string;
-};
-export type ApiError = {
-    status: 'error';
-    error: ApiErrorDetail;
-};
-
-export const isApiError = (error: unknown): error is ApiError =>
-    typeof error === 'object' &&
-    error !== null &&
-    'status' in error &&
-    error.status === 'error';
-
-export enum LightdashMode {
-    DEFAULT = 'default',
-    DEMO = 'demo',
-    PR = 'pr',
-    CLOUD_BETA = 'cloud_beta',
-    DEV = 'development',
-}
-
 export const isLightdashMode = (x: string): x is LightdashMode =>
     Object.values<string>(LightdashMode).includes(x);
 
@@ -1024,206 +430,6 @@ export enum LightdashInstallType {
     HEROKU = 'heroku',
     UNKNOWN = 'unknown',
 }
-
-export type SentryConfig = {
-    backend: {
-        dsn: string;
-        securityReportUri: string;
-    };
-    frontend: {
-        dsn: string;
-    };
-    release: string;
-    environment: string;
-    tracesSampleRate: number;
-    profilesSampleRate: number;
-    anr: {
-        enabled: boolean;
-        timeout?: number;
-        captureStacktrace: boolean;
-    };
-};
-
-export type HealthState = {
-    healthy: boolean;
-    mode: LightdashMode;
-    version: string;
-    localDbtEnabled: boolean;
-    defaultProject?: DbtProjectConfig;
-    isAuthenticated: boolean;
-    requiresOrgRegistration: boolean;
-    hasEmailClient: boolean;
-    hasMicrosoftTeams: boolean;
-    isServiceAccountEnabled: boolean;
-    isOrganizationWarehouseCredentialsEnabled: boolean;
-    latest: {
-        version?: string;
-    };
-    rudder: {
-        writeKey: string | undefined;
-        dataPlaneUrl: string | undefined;
-    };
-    sentry: Pick<
-        SentryConfig,
-        | 'frontend'
-        | 'release'
-        | 'environment'
-        | 'tracesSampleRate'
-        | 'profilesSampleRate'
-    >;
-    auth: {
-        disablePasswordAuthentication: boolean;
-        google: {
-            oauth2ClientId: string | undefined;
-            loginPath: string;
-            googleDriveApiKey: string | undefined;
-            enabled: boolean;
-            enableGCloudADC: boolean;
-        };
-        okta: {
-            enabled: boolean;
-            loginPath: string;
-        };
-        oneLogin: {
-            enabled: boolean;
-            loginPath: string;
-        };
-        azuread: {
-            enabled: boolean;
-            loginPath: string;
-        };
-        oidc: {
-            enabled: boolean;
-            loginPath: string;
-        };
-        pat: {
-            maxExpirationTimeInDays: number | undefined;
-        };
-        snowflake: {
-            enabled: boolean;
-        };
-    };
-    posthog:
-        | {
-              projectApiKey: string;
-              feApiHost: string;
-              beApiHost: string;
-          }
-        | undefined;
-    siteUrl: string;
-    intercom: {
-        appId: string;
-        apiBase: string;
-    };
-    pylon: {
-        appId: string;
-        verificationHash?: string;
-    };
-    staticIp: string;
-    query: {
-        maxLimit: number;
-        defaultLimit: number;
-        csvCellsLimit: number;
-        maxPageSize: number;
-    };
-    pivotTable: {
-        maxColumnLimit: number;
-    };
-    hasSlack: boolean;
-    hasGithub: boolean;
-    hasGitlab: boolean;
-    hasHeadlessBrowser: boolean;
-    hasExtendedUsageAnalytics: boolean;
-    hasCacheAutocompleResults: boolean;
-    appearance: {
-        overrideColorPalette: string[] | undefined;
-        overrideColorPaletteName: string | undefined;
-    };
-    isCustomRolesEnabled: boolean;
-    embedding: {
-        enabled: boolean;
-        events:
-            | {
-                  enabled: boolean;
-                  rateLimiting: {
-                      maxEventsPerWindow: number;
-                      windowDurationMs: number;
-                  };
-                  allowedOrigins: string[];
-                  enablePostMessage: boolean;
-              }
-            | undefined;
-    };
-    ai: {
-        analyticsProjectUuid?: string;
-        analyticsDashboardUuid?: string;
-    };
-};
-
-export enum DBFieldTypes {
-    DIMENSION = 'dimension',
-    METRIC = 'metric',
-}
-
-export const sensitiveDbtCredentialsFieldNames = [
-    'personal_access_token',
-    'api_key',
-] as const;
-
-export const DbtProjectTypeLabels: Record<DbtProjectType, string> = {
-    [DbtProjectType.DBT]: 'dbt local server',
-    [DbtProjectType.DBT_CLOUD_IDE]: 'dbt cloud',
-    [DbtProjectType.GITHUB]: 'Github',
-    [DbtProjectType.GITLAB]: 'GitLab',
-    [DbtProjectType.BITBUCKET]: 'BitBucket',
-    [DbtProjectType.AZURE_DEVOPS]: 'Azure DevOps',
-    [DbtProjectType.NONE]: 'CLI',
-    [DbtProjectType.MANIFEST]: 'Manifest',
-};
-
-export enum CreateProjectTableConfiguration {
-    PROD = 'prod',
-    ALL = 'all',
-}
-
-export type CreateProject = Omit<
-    Project,
-    | 'projectUuid'
-    | 'organizationUuid'
-    | 'schedulerTimezone'
-    | 'createdByUserUuid'
-> & {
-    warehouseConnection: CreateWarehouseCredentials;
-    copyWarehouseConnectionFromUpstreamProject?: boolean;
-    tableConfiguration?: CreateProjectTableConfiguration;
-    copyContent?: boolean;
-};
-
-export type CreateProjectOptionalCredentials = Omit<
-    CreateProject,
-    'warehouseConnection'
-> & {
-    warehouseConnection?: CreateWarehouseCredentials;
-};
-
-export const hasWarehouseCredentials = (
-    createProject: CreateProjectOptionalCredentials,
-): createProject is CreateProjectOptionalCredentials & {
-    warehouseConnection: CreateWarehouseCredentials;
-} =>
-    !!createProject.warehouseConnection &&
-    Object.keys(createProject.warehouseConnection).length > 0;
-
-export type UpdateProject = Omit<
-    Project,
-    | 'projectUuid'
-    | 'organizationUuid'
-    | 'type'
-    | 'schedulerTimezone'
-    | 'createdByUserUuid'
-> & {
-    warehouseConnection: CreateWarehouseCredentials;
-};
 
 export const getResultValueArray = (
     rows: ResultRow[],
