+++ packages/common/src/pivot/index.ts
@@ -1,3 +1,4 @@
 export * from './derivePivotConfigFromChart';
+export * from './pivotConfig';
 export * from './pivotQueryResults';
 export * from './utils';
