+++ packages/common/tsconfig.cjs.json
@@ -4,7 +4,8 @@
         "module": "CommonJS",
         "outDir": "dist/cjs",
         "sourceMap": true,
-        "composite": false,
+        "composite": true,
+        "tsBuildInfoFile": "dist/cjs/.tsbuildinfo",
         "declaration": true,
         "declarationMap": true,
     },
