+++ packages/common/tsconfig.esm.json
@@ -4,7 +4,8 @@
         "module": "ESNext",
         "outDir": "dist/esm",
         "sourceMap": true,
-        "composite": false,
+        "composite": true,
+        "tsBuildInfoFile": "dist/esm/.tsbuildinfo",
         "declaration": true,
         "declarationMap": true,
     },
