+++ packages/common/tsconfig.types.json
@@ -2,6 +2,7 @@
     "extends": "./tsconfig.json",
     "compilerOptions": {
         "outDir": "dist/types",
+        "tsBuildInfoFile": "dist/types/.tsbuildinfo",
         "declaration": true,
         "emitDeclarationOnly": true,
         "declarationMap": true,
