+++ packages/warehouses/tsconfig.json
@@ -12,14 +12,14 @@
         "allowJs": false, /* Allow javascript files to be compiled. */
         // "checkJs": true,                             /* Report errors in .js files. */
         // "jsx": "preserve",                           /* Specify JSX code generation: 'preserve', 'react-native', 'react', 'react-jsx' or 'react-jsxdev'. */
-        // "declaration": true,                         /* Generates corresponding '.d.ts' file. */
-        // "declarationMap": true,                      /* Generates a sourcemap for each corresponding '.d.ts' file. */
+        "declaration": true /* Generates corresponding '.d.ts' file. */,
+        "declarationMap": true /* Generates a sourcemap for each corresponding '.d.ts' file. */,
         // "sourceMap": true,                           /* Generates corresponding '.map' file. */
         // "outFile": "./",                             /* Concatenate and emit output to single file. */
         "outDir": "dist",
         "rootDir": "src",
         "composite": true /* Enable project compilation */,
-        // "tsBuildInfoFile": "./",                     /* Specify file to store incremental compilation information */
+        "tsBuildInfoFile": "dist/.tsbuildinfo" /* Specify file to store incremental compilation information */,
         // "removeComments": true,                      /* Do not emit comments to output. */
         // "noEmit": true,                              /* Do not emit outputs. */
         // "importHelpers": true,                       /* Import emit helpers from 'tslib'. */
@@ -68,4 +68,9 @@
         "src/**/*.ts",
         "src/**/*.json"
     ],
+    "references": [
+        {
+            "path": "../common"
+        }
+    ]
 }
